Navier Stokes Ideas
Navier Stokes differences are differential equations that describes the flow of fluids. Derived partials are pursued to determine the velocity and pressure fields in a flow. They were named after Claude-Louis Navier and George Gabriel Stokes developed a set of reflections that would describe the motion of fluid substances such as liquids and gases. These forces state that changes in momentum and interference of a fluid particle are simply the product (result) of changes in pressure and viscous dissipative forces (similar to friction) present in the fluid. This viscous force originates in molecular interaction.

So I tryed to create a software script to quantify and measure how chaotic a fluid can be. Are initial ideas, and am using this to pratice my knowledge in Phyton.